Add a mostly working API server.
This comes with a whole lot of caveats.  Basically, it works for me in
my testing on one specific setup.  It currently only supports connecting
transverters to a coax switch, and that to the Beaglebone.  It uses
flask but wants to be run under uwsgi.  It doesn't support TLS, so the
username and password used for doing anything with the hardware gets
sent in the clear.

But, it does actually work and do something if you're willing to do
exactly what I am doing.  Over time, this will get a whole lot more
flexible.